+++ index.html
@@ -226,6 +226,6 @@
       </div>
     </section>
   </template>
- <script scr="js/main.js"></script>
+ <script scr="js/main.js" type="module"></script>
 </body>
 </html>
